Move UriContentListener to test utils
